% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_vecchia.R
\name{dag_vecchia_predict}
\alias{dag_vecchia_predict}
\title{Generate a Directed Acyclic Graph (DAG) for predictions}
\usage{
dag_vecchia_predict(coords, newcoords, m = 20)
}
\arguments{
\item{coords}{A numeric matrix or data frame (\eqn{n \times d}) of training spatial coordinates, where \eqn{n} 
is the number of locations and \eqn{d} is the spatial dimension (e.g., 2 for latitude and longitude).}

\item{newcoords}{A numeric matrix or data frame (\eqn{n_{o} \times d}) of testing spatial coordinates, where \eqn{n_o} 
is the number of locations and \eqn{d} is the spatial dimension (e.g., 2 for latitude and longitude).}

\item{m}{An integer specifying the number of nearest neighbors to consider when constructing the DAG. 
Default is 20.}
}
\value{
The DAG for predictions
}
\description{
Generate a Directed Acyclic Graph (DAG) for predictions
}
