% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Compute ESS for every entry of an MCMC sample of N x K matrices.}
\usage{
effective_sample_size(draws)
}
\arguments{
\item{draws}{Cube of dimension N x K x T (slice t = sample t).}
}
\value{
N x K matrix of effective sample sizes.
}
\description{
Compute ESS for every entry of an MCMC sample of N x K matrices.
}
