% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_iox.R
\name{rgpiox}
\alias{rgpiox}
\title{Simulate GP-IOX data}
\usage{
rgpiox(
  coords,
  Sigma,
  Theta,
  m = 15,
  matern = 1,
  silent = TRUE,
  num_threads = RhpcBLASctl::get_num_cores()
)
}
\arguments{
\item{coords}{A numeric matrix of spatial coordinates. 
Must not contain missing values (`NA`).}

\item{Sigma}{A numeric, square covariance matrix of dimension `q x q`.}

\item{Theta}{A numeric matrix of spatial parameters of dimension `3 x q`. 
The rows correspond to `phi`, `nu`, and `alpha`, respectively.}

\item{m}{A single positive integer specifying the number of nearest neighbors 
to condition on for the Vecchia approximation. Default is `15`.}

\item{matern}{A single numeric value indicating the covariance 
specification. Default is `1` for Matern.}

\item{silent}{Logical. If `FALSE`, prints progress messages to the console. 
Default is `TRUE`.}

\item{num_threads}{Integer. Number of OpenMP threads to use for the simulation. 
Defaults to the number of available cores via `RhpcBLASctl::get_num_cores()`.}
}
\value{
A numeric matrix of simulated values with dimensions `n x q`, ordered 
  to match the original input `coords`.
}
\description{
Simulates GP-IOX data over a set of coordinates.
}
