% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfact}
\alias{sfact}
\title{Scaling factor of IOX.}
\usage{
sfact(dag, S, theta, matern = 1L, n_threads = 1L)
}
\arguments{
\item{dag}{an object returned from `spiox::dag_vecchia` or similar}

\item{S}{a matrix of reference coordinates}

\item{theta}{a matrix of dimension (4,q) where each column are the marginal covariance parameters of the corresponding variable}

\item{matern}{an integer. Options are 0: power exponential, 1: matern, 2: wave}

\item{n_threads}{integer number of threads for multi-threaded operations}
}
\value{
A matrix of dimension (q,q) filled in its lower-triangular portion. The (i,j) element is the scaling factor for \eqn{C_{ij}}.
The IOX cross-covariance at zero distance will be \eqn{Sigma_{ij}} multiplied by this scaling factor.
}
\description{
This function computes the scaling factor used to compute the cross-covariance at zero distance for IOX.
}
\details{
The function is designed for computing the IOX scaling factor used to compute the cross-covariances \eqn{C_{ij}}.
}
