% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spiox_response}
\alias{spiox_response}
\title{Spatial Response Model using Gaussian Processes with IOX.}
\usage{
spiox_response(
  Y,
  X,
  coords,
  custom_dag,
  theta_opts,
  Sigma_start,
  Beta_start,
  mcmc = 1000L,
  print_every = 100L,
  matern = 1L,
  sample_iwish = TRUE,
  sample_mvr = TRUE,
  sample_theta_gibbs = FALSE,
  upd_theta_opts = TRUE,
  num_threads = 1L
)
}
\arguments{
\item{Y}{A numeric matrix (\eqn{n \times q}) of observed multivariate spatial responses, 
where \eqn{n} is the number of spatial locations and \eqn{q} is the number of response variables.}

\item{X}{A numeric matrix (\eqn{n \times p}) of predictors corresponding to the observed responses.}

\item{coords}{A numeric matrix (\eqn{n \times d}) of spatial coordinates, where \eqn{d} is the spatial dimension 
(e.g., 2 for latitude and longitude).}

\item{custom_dag}{An object returned from `spiox::dag_vecchia`}

\item{theta_opts}{A numeric matrix specifying options for the correlation parameters (\eqn{\theta}) 
used during MCMC updates. The way this is input determines how MCMC works. See details below.}

\item{Sigma_start}{A numeric matrix (\eqn{q \times q}) specifying the starting value for the IOX covariance matrix 
\eqn{\Sigma}.}

\item{Beta_start}{A numeric matrix (\eqn{p \times q}) specifying the starting values for the regression coefficients.}

\item{mcmc}{An integer specifying the number of MCMC iterations to perform.}

\item{print_every}{An integer specifying the frequency of progress updates during MCMC iterations. Default is 100.}

\item{matern}{An integer flag for enabling Matérn correlation functions for spatial dependence modeling. Default is 1.
Other options: 0=power exponential, 2=wave.}

\item{sample_iwish}{A logical value indicating whether to sample the covariance matrix (\eqn{\Sigma}) 
via Gibbs update from an Inverse Wishart prior. Default is TRUE.}

\item{sample_mvr}{A logical value indicating whether to sample multivariate regression coefficients 
via Gibbs update from a Normal prior. Default is TRUE.}

\item{sample_theta_gibbs}{A logical value indicating whether to enable Gibbs sampling for the correlation 
parameters (\eqn{\theta}). This should be set to TRUE to run "IOX Grid" or "IOX Cluster" from the paper, otherwise FALSE.}

\item{upd_theta_opts}{A logical value indicating whether to update the correlation parameter options (\eqn{\theta}) 
adaptively during MCMC iterations. This should be set to TRUE to run "IOX Full" or "IOX Cluster" from the paper, otherwise FALSE.
The update is performed jointly for the whole vector if q=3 or less; conditionally in blocks if q>3.}

\item{num_threads}{An integer specifying the number of threads for parallel computation. Default is 1.}
}
\value{
A list containing:
\item{Beta}{Array of dimension (p,q,mcmc) with posterior samples of the regression coefficients (\eqn{\beta}).}
\item{Sigma}{Array of dimension (q,q,mcmc) with posterior samples of the covariance matrix (\eqn{\Sigma}).}
\item{theta}{Array of dimension (4,q,mcmc) with posterior samples of the correlation parameters (\eqn{\theta}).}
\item{theta_which}{Cluster membership for "IOX Grid" and "IOX Cluster".}
\item{theta_opts}{Cluster options for "IOX Grid" and "IOX Cluster".}
\item{timings}{Breakdown of timings of the various MCMC operations (debugging).}
}
\description{
This function performs Bayesian inference for a spatial response model using a 
multivariate GP with Inside-Out Cross-Covariance (IOX).
}
\details{
The function is designed for scalable inference on spatial multivariate data using GP-IOX. 
Use multi-threading (`num_threads > 1`) for faster computation on large datasets.
}
